requirements.txt
streamlit
gspread
google-auth
google-auth-oauthlib
pytesseract
Pillow